feat: surface agent dashboards on agent home page
Extract DashboardCard into a shared component and render it in the
agent home right column. Enrich the single-agent API endpoint so
dashboards data is actually returned (previously only the list
endpoint enriched). Card shows the screenshot with an "Open app"
outline button on hover and expands in-flow on hover.
