+++ src/renderer/components/agents/agent-home/agent-home.tsx
@@ -24,6 +24,7 @@ import { HomeExtras } from './home-extras'
 import { HomeConnections } from './home-connections'
 import { HomeVolumes } from './home-volumes'
 import { HomeBookmarks } from './home-bookmarks'
+import { DashboardCard } from '@renderer/components/home/dashboard-card'
 import { useUpdateAgent, useDeleteAgent, type ApiAgent } from '@renderer/hooks/use-agents'
 import { AgentCreationAids, type ImportResult } from '@renderer/components/agents/agent-creation-aids'
 import { useStartOnboardingSession } from '@renderer/hooks/use-start-onboarding-session'
@@ -481,6 +482,13 @@ export function AgentHome({ agent, onSessionCreated, onOpenSettings }: AgentHome
         {/* Right Column — Triggers + Connections + Skills + Volumes */}
         {showRightColumn && (
           <div className="space-y-3">
+            {(agent.dashboards ?? []).map((d) => (
+              <DashboardCard
+                key={d.slug}
+                dashboard={d}
+                agentSlug={agent.slug}
+              />
+            ))}
             <HomeTriggers
               agentSlug={agent.slug}
               scheduledTasks={scheduledTasks}
